move command related functions to cmd.go
